test(dql): Split mutation test (#20)
Description: Splits apart the mutation_test.go in dgraph/worker into
unit tests vs. integration tests. The purpose is to allow for more
effective local tests.
