+++ worker/mutation_integration_test.go
@@ -19,7 +19,6 @@
 package worker
 
 import (
-	"reflect"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -27,90 +26,8 @@ import (
 	"github.com/dgraph-io/dgraph/posting"
 	"github.com/dgraph-io/dgraph/protos/pb"
 	"github.com/dgraph-io/dgraph/schema"
-	"github.com/dgraph-io/dgraph/types"
 	"github.com/dgraph-io/dgraph/x"
 )
-
-func TestConvertEdgeType(t *testing.T) {
-	var testEdges = []struct {
-		input     *pb.DirectedEdge
-		to        types.TypeID
-		expectErr bool
-		output    *pb.DirectedEdge
-	}{
-		{
-			input: &pb.DirectedEdge{
-				Value: []byte("set edge"),
-				Attr:  x.GalaxyAttr("name"),
-			},
-			to:        types.StringID,
-			expectErr: false,
-			output: &pb.DirectedEdge{
-				Value:     []byte("set edge"),
-				Attr:      x.GalaxyAttr("name"),
-				ValueType: 9,
-			},
-		},
-		{
-			input: &pb.DirectedEdge{
-				Value: []byte("set edge"),
-				Attr:  x.NamespaceAttr(0xf2, "name"),
-				Op:    pb.DirectedEdge_DEL,
-			},
-			to:        types.StringID,
-			expectErr: false,
-			output: &pb.DirectedEdge{
-				Value:     []byte("set edge"),
-				Attr:      x.NamespaceAttr(0xf2, "name"),
-				Op:        pb.DirectedEdge_DEL,
-				ValueType: 9,
-			},
-		},
-		{
-			input: &pb.DirectedEdge{
-				ValueId: 123,
-				Attr:    x.GalaxyAttr("name"),
-			},
-			to:        types.StringID,
-			expectErr: true,
-		},
-		{
-			input: &pb.DirectedEdge{
-				Value: []byte("set edge"),
-				Attr:  x.GalaxyAttr("name"),
-			},
-			to:        types.UidID,
-			expectErr: true,
-		},
-	}
-
-	for _, testEdge := range testEdges {
-		err := ValidateAndConvert(testEdge.input,
-			&pb.SchemaUpdate{
-				ValueType: pb.Posting_ValType(testEdge.to),
-			})
-		if testEdge.expectErr {
-			require.Error(t, err)
-		} else {
-			require.NoError(t, err)
-			require.True(t, reflect.DeepEqual(testEdge.input, testEdge.output))
-		}
-	}
-
-}
-
-func TestValidateEdgeTypeError(t *testing.T) {
-	edge := &pb.DirectedEdge{
-		Value: []byte("set edge"),
-		Attr:  x.GalaxyAttr("name"),
-	}
-
-	err := ValidateAndConvert(edge,
-		&pb.SchemaUpdate{
-			ValueType: pb.Posting_ValType(types.DateTimeID),
-		})
-	require.Error(t, err)
-}
 
 func TestPopulateMutationMap(t *testing.T) {
 	edges := []*pb.DirectedEdge{{
@@ -213,57 +130,4 @@ func TestCheckSchema(t *testing.T) {
 	require.NoError(t, err)
 	require.NoError(t, checkSchema(result.Preds[0]))
 	require.NoError(t, checkSchema(result.Preds[1]))
-}
-
-func TestTypeSanityCheck(t *testing.T) {
-	// Empty field name check.
-	typeDef := &pb.TypeUpdate{
-		Fields: []*pb.SchemaUpdate{
-			{
-				Predicate: x.GalaxyAttr(""),
-			},
-		},
-	}
-	err := typeSanityCheck(typeDef)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "Field in type definition must have a name")
-
-	// Object type without object name.
-	typeDef = &pb.TypeUpdate{
-		Fields: []*pb.SchemaUpdate{
-			{
-				Predicate: x.GalaxyAttr("name"),
-				ValueType: pb.Posting_OBJECT,
-			},
-		},
-	}
-	err = typeSanityCheck(typeDef)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "Field with value type OBJECT must specify the name")
-
-	// Field with directive.
-	typeDef = &pb.TypeUpdate{
-		Fields: []*pb.SchemaUpdate{
-			{
-				Predicate: x.GalaxyAttr("name"),
-				Directive: pb.SchemaUpdate_REVERSE,
-			},
-		},
-	}
-	err = typeSanityCheck(typeDef)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "Field in type definition cannot have a directive")
-
-	// Field with tokenizer.
-	typeDef = &pb.TypeUpdate{
-		Fields: []*pb.SchemaUpdate{
-			{
-				Predicate: x.GalaxyAttr("name"),
-				Tokenizer: []string{"int"},
-			},
-		},
-	}
-	err = typeSanityCheck(typeDef)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "Field in type definition cannot have tokenizers")
 }
